# docs/swagger.yaml
definitions:
...
  server.ApiResponse:
    properties:
      code:
        description: 状态码
        type: integer
      message:
        description: 状态短语
        type: string
      result:
        description: 数据结果集
        type: object
    type: object
...
info:
  contact: {}
paths:
...
  /api/file/upload:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: Bearer +token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文件
        in: body
        name: file[]
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  items:
                    type: string
                  type: array
              type: object
      summary: 上传文件
      tags:
      - api/file
...
  /file/upload:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 图片编码格式(e.g.image/jpeg)
        in: header
        name: X-content-type
        required: true
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 文件
        in: body
        name: file[]
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  items:
                    type: string
                  type: array
              type: object
      summary: 上传文件
      tags:
      - file
...
swagger: "2.0"